Extract download routes + bundle/overlay-image helpers (PR G) (#1329)
